requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Databricks
databricks-sql-connector>=3.0.0
databricks-sdk>=0.20.0

# LLM
anthropic>=0.18.0

# UI
gradio>=4.15.0

# SQL Validation
sqlglot>=20.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
black>=24.0.0
ruff>=0.1.0

# Utilities
rich>=13.0.0
tabulate>=0.9.0